Moved Mediapart related files to own bundle
