+++ src/BD/SiteReaderProxy/MediapartSiteBundle/Resources/config/services.yml
@@ -1,7 +1,13 @@
 parameters:
-#    bd_site_reader_proxy_mediapart_site.example.class: BD\SiteReaderProxy\MediapartSiteBundle\Example
+    bd_site_reader_proxy.mediapart_browser.class: BD\SiteReaderProxy\MediapartSiteBundle\MediapartBrowser
+    bd_site_reader_proxy.controller.mediapart.class: BD\SiteReaderProxy\MediapartSiteBundle\Controller\MediapartController
 
 services:
-#    bd_site_reader_proxy_mediapart_site.example:
-#        class: %bd_site_reader_proxy_mediapart_site.example.class%
-#        arguments: [@service_id, "plain_value", %parameter%]
+    bd_mediapart_site_reader.controller.mediapart:
+        class: %bd_site_reader_proxy.controller.mediapart.class%
+        arguments: [@bd_mediapart_site_reader.mediapart_browser]
+
+    bd_mediapart_site_reader.mediapart_browser:
+        class: %bd_site_reader_proxy.mediapart_browser.class%
+        arguments: [%mediapart_login%, %mediapart_password%, %app_uri%, %mediapart_session_cookie_string%]
+
