+++ src/BD/SiteReaderProxyBundle/Resources/config/services.yml
@@ -1,17 +1,7 @@
 parameters:
-    bd_site_reader_proxy.mediapart_browser.class: BD\SiteReaderProxyBundle\Proxy\MediapartBrowser
-    bd_site_reader_proxy.controller.feed_proxy.class: BD\SiteReaderProxyBundle\Controller\FeedProxyController
     bd_site_reader_proxy.routing_loader.yaml_finder.factory_class: BD\SiteReaderProxyBundle\DependencyInjection\Factory\YamlSitesRoutesFinderFactory
 
 services:
-    bd_site_reader_proxy.controller.feed_proxy:
-        class: %bd_site_reader_proxy.controller.feed_proxy.class%
-        arguments: [@bd_site_reader_proxy.mediapart_browser]
-
-    bd_site_reader_proxy.mediapart_browser:
-        class: %bd_site_reader_proxy.mediapart_browser.class%
-        arguments: [%mediapart_login%, %mediapart_password%, %app_uri%, %mediapart_session_cookie_string%]
-
     bd_site_reader_proxy.routing_loader:
         class: BD\SiteReaderProxyBundle\Routing\SiteRoutingLoader
         tags:
